comment service, handler

+++ post-service/internal/model/comment.go
@@ -5,5 +5,5 @@ import "gorm.io/gorm"
 type Comment struct {
 	gorm.Model
 	Content string `json:"content"`
-	PostID  uint64 `json:"postId"`
+	PostID  uint   `json:"postId"`
 }
